Use IsReady interface instead of the override function
