+++ platform-operator/controllers/verrazzano/component/authproxy/authproxy_component_test.go
@@ -7,79 +7,12 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	globalconst "github.com/verrazzano/verrazzano/pkg/constants"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
-	appsv1 "k8s.io/api/apps/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
 const profilesRelativePath = "../../../../manifests/profiles"
-
-// TestIsReady tests the AuthProxy IsReady call
-// GIVEN a AuthProxy component
-//  WHEN I call IsReady when all requirements are met
-//  THEN true or false is returned
-func TestIsReady(t *testing.T) {
-	tests := []struct {
-		name       string
-		client     client.Client
-		expectTrue bool
-	}{
-		{
-			name: "Test IsReady when AuthProxy is successfully deployed",
-			client: fake.NewFakeClientWithScheme(testScheme,
-				&appsv1.Deployment{
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: globalconst.VerrazzanoSystemNamespace,
-						Name:      "verrazzano-authproxy",
-					},
-					Status: appsv1.DeploymentStatus{
-						Replicas:            1,
-						ReadyReplicas:       1,
-						AvailableReplicas:   1,
-						UnavailableReplicas: 0,
-					},
-				}),
-			expectTrue: true,
-		},
-		{
-			name: "Test IsReady when AuthProxy deployment is not ready",
-			client: fake.NewFakeClientWithScheme(testScheme,
-				&appsv1.Deployment{
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: globalconst.VerrazzanoSystemNamespace,
-						Name:      "verrazzano-authproxy",
-					},
-					Status: appsv1.DeploymentStatus{
-						Replicas:            1,
-						ReadyReplicas:       1,
-						AvailableReplicas:   0,
-						UnavailableReplicas: 1,
-					},
-				}),
-			expectTrue: false,
-		},
-		{
-			name:       "Test IsReady when AuthProxy deployment does not exist",
-			client:     fake.NewFakeClientWithScheme(testScheme),
-			expectTrue: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := spi.NewFakeContext(tt.client, &vzapi.Verrazzano{}, false)
-			if tt.expectTrue {
-				assert.True(t, NewComponent().IsReady(ctx))
-			} else {
-				assert.False(t, NewComponent().IsReady(ctx))
-			}
-		})
-	}
-}
 
 // TestIsEnabled tests the AuthProxy IsEnabled call
 // GIVEN a AuthProxy component
