+++ platform-operator/controllers/verrazzano/component/coherence/coherence_test.go
@@ -1,29 +1,27 @@
-// Copyright (c) 2021, Oracle and/or its affiliates.
+// Copyright (c) 2021, 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 package coherence
 
 import (
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/verrazzano/verrazzano/platform-operator/constants"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8scheme "k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
-// TestIsCoherenceOperatorReady tests the IsApplicationOperatorReady function
-// GIVEN a call to IsCoherenceOperatorReady
+// TestIsCoherenceOperatorReady tests the isCoherenceOperatorReady function
+// GIVEN a call to isCoherenceOperatorReady
 //  WHEN the deployment object has enough replicas available
 //  THEN true is returned
 func TestIsCoherenceOperatorReady(t *testing.T) {
-
 	fakeClient := fake.NewFakeClientWithScheme(k8scheme.Scheme, &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Namespace: constants.VerrazzanoSystemNamespace,
-			Name:      coherenceOperatorDeploymentName,
+			Namespace: ComponentNamespace,
+			Name:      ComponentName,
 		},
 		Status: appsv1.DeploymentStatus{
 			Replicas:            1,
@@ -32,19 +30,18 @@ func TestIsCoherenceOperatorReady(t *testing.T) {
 			UnavailableReplicas: 0,
 		},
 	})
-	assert.True(t, IsCoherenceOperatorReady(spi.NewFakeContext(fakeClient, nil, false), "", constants.VerrazzanoSystemNamespace))
+	assert.True(t, isCoherenceOperatorReady(spi.NewFakeContext(fakeClient, nil, false)))
 }
 
-// TestIsCoherenceOperatorNotReady tests the IsApplicationOperatorReady function
-// GIVEN a call to IsCoherenceOperatorReady
+// TestIsCoherenceOperatorNotReady tests the isCoherenceOperatorReady function
+// GIVEN a call to isCoherenceOperatorReady
 //  WHEN the deployment object does NOT have enough replicas available
 //  THEN false is returned
 func TestIsCoherenceOperatorNotReady(t *testing.T) {
-
 	fakeClient := fake.NewFakeClientWithScheme(k8scheme.Scheme, &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Namespace: constants.VerrazzanoSystemNamespace,
-			Name:      coherenceOperatorDeploymentName,
+			Namespace: ComponentNamespace,
+			Name:      ComponentName,
 		},
 		Status: appsv1.DeploymentStatus{
 			Replicas:            1,
@@ -53,5 +50,5 @@ func TestIsCoherenceOperatorNotReady(t *testing.T) {
 			UnavailableReplicas: 1,
 		},
 	})
-	assert.False(t, IsCoherenceOperatorReady(spi.NewFakeContext(fakeClient, nil, false), "", constants.VerrazzanoSystemNamespace))
+	assert.False(t, isCoherenceOperatorReady(spi.NewFakeContext(fakeClient, nil, false)))
 }
