+++ platform-operator/controllers/verrazzano/component/kiali/kiali_test.go
@@ -1,15 +1,18 @@
-// Copyright (c) 2021, Oracle and/or its affiliates.
+// Copyright (c) 2021, 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 package kiali
 
 import (
 	"fmt"
-	"github.com/verrazzano/verrazzano/pkg/bom"
-	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/verrazzano/verrazzano/pkg/bom"
+	"github.com/verrazzano/verrazzano/pkg/helm"
+	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
+	appsv1 "k8s.io/api/apps/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8scheme "k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
@@ -36,4 +39,69 @@ func TestAppendOverrides(t *testing.T) {
 	assert.Len(t, kvs, 2)
 	assert.Equal(t, bom.KeyValue{Key: "key1", Value: "value1"}, kvs[0])
 	assert.Equal(t, bom.KeyValue{Key: webFQDNKey, Value: fmt.Sprintf("%s.default.mydomain.com", kialiHostName)}, kvs[1])
+}
+
+// TestIsKialiReady tests the isKialiReady function
+// GIVEN a call to isKialiReady
+//  WHEN the deployment object has enough replicas available
+//  THEN true is returned
+func TestIsKialiReady(t *testing.T) {
+	helm.SetChartStatusFunction(func(releaseName string, namespace string) (string, error) {
+		return helm.ChartStatusDeployed, nil
+	})
+	defer helm.SetDefaultChartStatusFunction()
+
+	fakeClient := fake.NewFakeClientWithScheme(testScheme, &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: ComponentNamespace,
+			Name:      kialiSystemName,
+		},
+		Status: appsv1.DeploymentStatus{
+			Replicas:            1,
+			ReadyReplicas:       1,
+			AvailableReplicas:   1,
+			UnavailableReplicas: 0,
+		},
+	})
+
+	assert.True(t, isKialiReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false)))
+}
+
+// TestIsKialiNotReady tests the isKialiReady function
+// GIVEN a call to isKialiReady
+//  WHEN the deployment object does NOT have enough replicas available
+//  THEN false is returned
+func TestIsKialiNotReady(t *testing.T) {
+	helm.SetChartStatusFunction(func(releaseName string, namespace string) (string, error) {
+		return helm.ChartStatusDeployed, nil
+	})
+	defer helm.SetDefaultChartStatusFunction()
+
+	fakeClient := fake.NewFakeClientWithScheme(testScheme, &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: ComponentNamespace,
+			Name:      kialiSystemName,
+		},
+		Status: appsv1.DeploymentStatus{
+			Replicas:            1,
+			ReadyReplicas:       0,
+			AvailableReplicas:   0,
+			UnavailableReplicas: 1,
+		},
+	})
+	assert.False(t, isKialiReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false)))
+}
+
+// TestIsKialiNotReadyChartNotFound tests the isKialiReady function
+// GIVEN a call to isKialiReady
+//  WHEN the Kiali chart is not found
+//  THEN false is returned
+func TestIsKialiNotReadyChartNotFound(t *testing.T) {
+	helm.SetChartStatusFunction(func(releaseName string, namespace string) (string, error) {
+		return helm.ChartNotFound, nil
+	})
+	defer helm.SetDefaultChartStatusFunction()
+
+	fakeClient := fake.NewFakeClientWithScheme(testScheme)
+	assert.False(t, isKialiReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false)))
 }
