+++ platform-operator/controllers/verrazzano/component/nginx/nginx.go
@@ -23,9 +23,6 @@ import (
 )
 
 const (
-	// ComponentNamespace is the NGINX namespace for verrazzano
-	ComponentNamespace = "ingress-nginx"
-
 	// ValuesFileOverride Name of the values file override for NGINX
 	ValuesFileOverride = "ingress-nginx-values.yaml"
 
@@ -33,10 +30,10 @@ const (
 	backendName    = "ingress-controller-ingress-nginx-defaultbackend"
 )
 
-func IsReady(context spi.ComponentContext, name string, namespace string) bool {
+func isNginxReady(context spi.ComponentContext) bool {
 	deployments := []types.NamespacedName{
-		{Name: ControllerName, Namespace: namespace},
-		{Name: backendName, Namespace: namespace},
+		{Name: ControllerName, Namespace: ComponentNamespace},
+		{Name: backendName, Namespace: ComponentNamespace},
 	}
 	prefix := fmt.Sprintf("Component %s", context.GetComponent())
 	return status.DeploymentsReady(context.Log(), context.Client(), deployments, 1, prefix)
