+++ platform-operator/controllers/verrazzano/component/oam/oam_test.go
@@ -3,12 +3,11 @@
 package oam
 
 import (
-	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/verrazzano/verrazzano/platform-operator/constants"
+	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
+	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8scheme "k8s.io/client-go/kubernetes/scheme"
@@ -27,15 +26,15 @@ var crEnabled = vzapi.Verrazzano{
 	},
 }
 
-// TestIsOAMOperatorReady tests the IsOAMReady function
-// GIVEN a call to IsOAMReady
+// TestIsOAMOperatorReady tests the isOAMReady function
+// GIVEN a call to isOAMReady
 //  WHEN the deployment object has enough replicas available
 //  THEN true is returned
 func TestIsOAMOperatorReady(t *testing.T) {
 
 	fakeClient := fake.NewFakeClientWithScheme(k8scheme.Scheme, &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Namespace: constants.VerrazzanoSystemNamespace,
+			Namespace: ComponentNamespace,
 			Name:      oamOperatorDeploymentName,
 		},
 		Status: appsv1.DeploymentStatus{
@@ -45,18 +44,18 @@ func TestIsOAMOperatorReady(t *testing.T) {
 			UnavailableReplicas: 0,
 		},
 	})
-	assert.True(t, IsOAMReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false), "", constants.VerrazzanoSystemNamespace))
+	assert.True(t, isOAMReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false)))
 }
 
-// TestIsOAMOperatorNotReady tests the IsOAMReady function
-// GIVEN a call to IsOAMReady
+// TestIsOAMOperatorNotReady tests the isOAMReady function
+// GIVEN a call to isOAMReady
 //  WHEN the deployment object does NOT have enough replicas available
 //  THEN false is returned
 func TestIsOAMOperatorNotReady(t *testing.T) {
 
 	fakeClient := fake.NewFakeClientWithScheme(k8scheme.Scheme, &appsv1.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
-			Namespace: constants.VerrazzanoSystemNamespace,
+			Namespace: ComponentNamespace,
 			Name:      oamOperatorDeploymentName,
 		},
 		Status: appsv1.DeploymentStatus{
@@ -66,7 +65,7 @@ func TestIsOAMOperatorNotReady(t *testing.T) {
 			UnavailableReplicas: 1,
 		},
 	})
-	assert.False(t, IsOAMReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false), "", constants.VerrazzanoSystemNamespace))
+	assert.False(t, isOAMReady(spi.NewFakeContext(fakeClient, &vzapi.Verrazzano{}, false)))
 }
 
 // TestIsEnabledNilOAM tests the IsEnabled function
