+++ platform-operator/controllers/verrazzano/component/verrazzano/verrazzano_component.go
@@ -4,20 +4,23 @@
 package verrazzano
 
 import (
-	"fmt"
 	"path/filepath"
 
-	globalconst "github.com/verrazzano/verrazzano/pkg/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/helm"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/istio"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/nginx"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
-	"github.com/verrazzano/verrazzano/platform-operator/internal/k8s/status"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/vzconfig"
 	"k8s.io/apimachinery/pkg/types"
 )
+
+// ComponentName is the name of the component
+const ComponentName = "verrazzano"
+
+// ComponentNamespace is the namespace of the component
+const ComponentNamespace = constants.VerrazzanoSystemNamespace
 
 type verrazzanoComponent struct {
 	helm.HelmComponent
@@ -30,7 +33,7 @@ func NewComponent() spi.Component {
 		helm.HelmComponent{
 			ReleaseName:             ComponentName,
 			ChartDir:                filepath.Join(config.GetHelmChartsDir(), ComponentName),
-			ChartNamespace:          constants.VerrazzanoSystemNamespace,
+			ChartNamespace:          ComponentNamespace,
 			IgnoreNamespaceOverride: true,
 			ResolveNamespaceFunc:    resolveVerrazzanoNamespace,
 			AppendOverridesFunc:     appendVerrazzanoOverrides,
@@ -60,23 +63,12 @@ func (c verrazzanoComponent) PreUpgrade(ctx spi.ComponentContext) error {
 		c.ReleaseName, resolveVerrazzanoNamespace(c.ChartNamespace), c.ChartDir)
 }
 
-// IsReady Verrazzano component ready-check
+// IsReady component check
 func (c verrazzanoComponent) IsReady(ctx spi.ComponentContext) bool {
-	if !c.HelmComponent.IsReady(ctx) {
-		return false
+	if c.HelmComponent.IsReady(ctx) {
+		return isVerrazzanoReady(ctx)
 	}
-	var deployments []types.NamespacedName
-	if isVMOEnabled(ctx.EffectiveCR()) {
-		deployments = append(deployments, []types.NamespacedName{
-			{Name: "verrazzano-operator", Namespace: globalconst.VerrazzanoSystemNamespace},
-			{Name: "verrazzano-monitoring-operator", Namespace: globalconst.VerrazzanoSystemNamespace},
-		}...)
-	}
-	prefix := fmt.Sprintf("Component %s", ctx.GetComponent())
-	if !status.DeploymentsReady(ctx.Log(), ctx.Client(), deployments, 1, prefix) {
-		return false
-	}
-	return isVerrazzanoSecretReady(ctx)
+	return false
 }
 
 // PostInstall - post-install, clean up temp files
@@ -120,28 +112,28 @@ func (c verrazzanoComponent) GetIngressNames(ctx spi.ComponentContext) []types.N
 
 	if vzconfig.IsElasticsearchEnabled(ctx.EffectiveCR()) {
 		ingressNames = append(ingressNames, types.NamespacedName{
-			Namespace: constants.VerrazzanoSystemNamespace,
+			Namespace: ComponentNamespace,
 			Name:      constants.ElasticsearchIngress,
 		})
 	}
 
 	if vzconfig.IsGrafanaEnabled(ctx.EffectiveCR()) {
 		ingressNames = append(ingressNames, types.NamespacedName{
-			Namespace: constants.VerrazzanoSystemNamespace,
+			Namespace: ComponentNamespace,
 			Name:      constants.GrafanaIngress,
 		})
 	}
 
 	if vzconfig.IsKibanaEnabled(ctx.EffectiveCR()) {
 		ingressNames = append(ingressNames, types.NamespacedName{
-			Namespace: constants.VerrazzanoSystemNamespace,
+			Namespace: ComponentNamespace,
 			Name:      constants.KibanaIngress,
 		})
 	}
 
 	if vzconfig.IsPrometheusEnabled(ctx.EffectiveCR()) {
 		ingressNames = append(ingressNames, types.NamespacedName{
-			Namespace: constants.VerrazzanoSystemNamespace,
+			Namespace: ComponentNamespace,
 			Name:      constants.PrometheusIngress,
 		})
 	}
